Factor out cookie handling into its own file.
